+++ vae/main.py
@@ -9,6 +9,7 @@ from helper_plotting import *
 from torch.utils.data import Dataset, DataLoader
 import os
 import cv2
+from dataset import CelebADataset
 ##########################
 ### SETTINGS
 ##########################
@@ -44,51 +45,7 @@ custom_transforms = torchvision.transforms.Compose([
 ])
 
 
-
-
 dataroot = './data/celeba/'
-
-
-class CelebADataset(Dataset):
-  
-    def __init__(self,  root_dir,
-                        attrs,
-                        filenames,
-                        transforms=None):
-        '''
-         initilaize the important passed variables
-        '''
-        self.root_dir=root_dir
-        self.transforms=transforms
-        self.attrs = attrs
-        self.filenames = filenames
-        
-    def __len__(self):
-        '''
-        returns the no of dataset
-        '''
-        return len(self.filenames)
-    
-    def __getitem__(self, index):
-        '''
-        returns image in numpy with labels
-        '''
-
-        # data/celeba/
-        img_path = os.path.join(self.root_dir, 'img_align_celeba', self.filenames[index])
-        image = cv2.imread(img_path)
-        image = cv2.cvtColor(image, cv2.COLOR_BGR2RGB)
-
-        # select <index> row
-        target = np.array(self.attrs[index,:])
-        target[target==-1] = 0
-    
-        if self.transforms:
-          image = self.transforms(image)
-
-        return (image,target)
-
-
 
 dataset = CelebADataset(root_dir=dataroot,
                         attrs=attrs,
@@ -118,7 +75,7 @@ test_loader = torch.utils.data.DataLoader(test_dataset, batch_size=BATCH_SIZE,
 
 
 # Image batch dimensions: torch.Size([5000, 3, 128, 128])
-# Image label dimensions: torch.Size([5000, 39])
+# Image label dimensions: torch.Size([5000, 40])
 torch.manual_seed(RANDOM_SEED)
 for idx, (images, labels) in enumerate(train_loader):
     # labels = lab[]
